Move makeScript helper
