+++ front/scripts/helpers.ts
